Warning for BattleDialog before EncounterStarting
If you place BattleDialog outside of EncounterStarting ("outside of all
functions"), then it would be called before the encounter has started.

This causes issues because the UI controller doesn't actually exist yet.

With this commit, any BattleDialog calls from before the encounter has
started will be ignored with a warning.