Fixed encounter storage for good
The script object has a property called 'instances', which is a list of
every script loaded by the engine.

That's fine, but there's a problem: it never empties when a battle ends.
That means that every time you open a mod, every single script in its
Encounters, Monsters and Waves folders will be stored in memory for the
entire rest of the session.

I have fixed this here. This should hopefully fix some memory usage,
especially when opening mods that have a lot of waves.