Included Arial as a font
So far, the font Arial has never been included in CYF, and was instead
always taken from the OS. This means that since Mac and Linux do not
have this font, they would not see any text that used this font.

With this commit, this font has been added, and the scenes that use
Arial now use its locally imported version.