Big main menu overhaul
* Simplified Disclaimer screen

* Much nicer looking mod selection screen that allows previews, and
automatically brings you back to your last played mod. It will not
display mods with 0 encounters, or mods in Hidden folders

* Fix ESCAPE reloading ModSelect after exiting a battle. This bug has
existed since Unitale. Now, pressing escape on ModSelect will return you
to the Disclaimer screen (useful if you wanted to open the overworld,
but tried to focus the window first by clicking, and it opened ModSelect
instead)

* Fixed 'inBattle' being set improperly

* Better Crate Your Frisk text for ModSelect

* The new Mod Selection screen may potentially take longer to load if
you have a lot of mods, because it needs to search each mod for
encounter files as well. A "Loading Mods" message has been added to the
Disclaimer screen for whenever you click.