Replaced ESC on ModSelect functionality with button
* Removed the ability to exit the Mod Select screen by pressing Escape.
If the player was holding Escape while exiting a battle, they would be
taken straight to the Disclaimer screen. This proved to be unsolvable,
and is actually because of a Unity bug.

* Added an Exit button to the Mod Select screen. It will fade out to
half opacity after a few seconds, or immediately if you have just exited
a mod. You can mouse over it to make it visible again, and of course,
click it to go to the Disclaimer screen.