Inventory.AddItem can now take an index
* Inventory.AddItem(name) can now optionally be Inventory.AddItem(name,
index=8). The new item will be added at the given index.

If the index is < the number of items in the inventory, the new item
will appear at the proper slot and "push away" the item that was there.

If the index is > the number of items in the inventory, it will be added
to the end.