Bug fix: Can't move player on first frame of waves
There has always been a strange bug in CYF where you were unable to move
the player on the first frame of a wave. I found out why: there is some
strange code in PlayerController.cs that wasn't in Unitale's code, and
it's to blame. This fix comments it out, and has been tested and
confirmed working!