Properly throw an error for non-existing Audio
* Made CYF show an error when the user tries to load an Audio clip that
doesn't exist. Now both audio and sprites will error like this, instead
of just sprites!

This is disabled in retromode, although I don't know why you would want
to disable this error message, as it would just print warnings to the
debugger, which is about the same as just erroring.