Force windowed mode on first launch
Solves #66 .

Here's how it works:

Unity games - all of them - store a series of keys and values to what's called the "PlayerPrefs". You can check their paths here: https://docs.unity3d.com/ScriptReference/PlayerPrefs.html

If PlayerPrefs do not already exist, that means the game has not yet been opened on this machine. The PlayerPrefs automatically get created. And when this happens with CYF, it tries to use a default screen resolution - which ends up opening the game in fullscreen and, on some machines, will fall to a resolution of 1x1 pixels.

This commit forces the first ever launch of the engine (using PlayerPrefs) to force the game to windowed mode, and explicitly give it a numbered resolution, multiple of 640x480. Note that it will also happen for people opening CYF v0.6.4.1 for the first time - but, again, it's only one time; after this initial one time, it'll go back to the standard behavior of storing the user's last resolution.