Modify [font:x] CYFException
* Fix no error screen showing, leading to engine freezes and invisible enemy text bubbles, when trying to load an invalid font in some cases

~~~~~

Up to now, when an invalid font was switched to using [font:x] at any point, the program would freeze, because an error would be thrown but the typical dog error screen would not show. This also resulted in side effects like enemy speech bubbles not appearing.

Now, no matter what the source is, an error message will always be shown when [font:x] is used to switch to an invalid font. This will now also affect things like using [font:x] in a monster's name, an item name, or in encountertext or enemy comments.

Sadly, this means that for the previously-covered cases, such as using [font:x] inside of CreateText, you will no longer see the line number - but you will still see the name of the attempted font in the message. This is unfortunately a necessary sacrifice.