[lettereffect] skipping + [lettereffect:twitch]
* Fixed `[lettereffect]` not applying to letters that get skipped over by the player, with the `[instant]` and `[instant:allowcommand]` commands, and `text.SkipLine()`
* Fixed `[lettereffect]`'s version of "twitch" taking an absurd 5 seconds (at minimum) for each letter to start twitching, per letter

note: on line 603, I had to comment out a line with `InUpdateControlCommand`. there's a good reason for this. I found out that on line 641, the exact same command gets called again. there were issues when calling an `instant` command twice in one frame like this when using the new code that invokes `HandleShowLetter` when `InUpdateControlCommand` is called with an `instant` command, so I made only one of them get called in the end.

it's possible this commit may make some more issues. this entire file is honestly a mess. I sure hope it doesn't, but this commit is probably the first thing to check if there are any text command related issues after this point