FIX: All [func] commands used in the Encounter text call functions from the Encounter
Removes the inconsistency related to the main text in the arena when it
comes to which script to pick functions from when using the text command
[func], it now always tries to fetch a function from the Encounter
script.

The function UIController.ActionDialogResult() has also been simplified
since its third argument is never used, and his second argument has been
given a default value.