FIX: BattleDialog() checks its data and returns a better error
BattleDialog() used to only check whether it has a table or a string as
its argument, but it didn't check the contents of the table in the
first case.

On top of that, if you were to pass something other than a table or a
string as BattleDialog()'s argument, the error used to not tell the
modder about the erroneous line.