Add enemy management (#85)
* Add CreateEnemy(enemy,x,y)

This PR adds the encounter function CreateEnemy, which lets
you create an enemy while the encounter is running. This
function returns the ScriptWrapper object, which is not
added to the default enemies table.

* Add Remove() to enemy scripts

This function removes the enemy from the battle. Note that
this does not actually remove the script reference; it
solely removes the GameObject (and removes it from the
internal enemies list.) It will throw an error if you
try to destroy an enemy which is already destroyed.