Added bullet.OnHit()
Now bullets have a OnHit variable which is a Lua function.
Said function works much like the standard OnHit function, which has a
argument named "bullet" which is the bullet itself.

If a function has a OnHit variable, then it will be run instead of the
script's general OnHit function.

The function and the bullet have to be created in the same script.