Change UIStats TextManagers to LuaTextManagers; implement UI.UpdateInfo()
This commit changes both the NameLv and HPText objects to instead be instances of LuaTextManager, and are
now accessible from the `UI` object.

`UI.UpdateInfo()` has also been added which currently repositions the HP bar depending on the number of
letters in the NameLv object. Additionally, this means the behavior of the repositioning of the HP bar now
depends on NameLv, and not only the name being > 6 characters.

A minor change to the Disclaimer scene was added which changes "v0.6.5" to "v0.6.6" solely because I'm
tired of looking at it. This is only in the Unity editor and does not change the string in the actual engine.