Fix CYF freezing and taking up all of the computer's memory if the sprite to dust using sprite.Dust() is too big
In order to add layering to particle systems, Unity added a neat way to
transform a particle system's particles into a mesh using its BakeMesh()
function.

However, if the particle system has more than exactly 16383 particles
(or 2^14 - 1), Unity inexplicably freezes and takes all of the
computer's memory over time.

This commit added a way to create slices of exactly 16383 particles and
create a particle system for each of them instead, allowing for the
dusting of bigger sprites without encountering this problem.