Added the variable "noscalerotationbug" to disable a bug
In CYF 0.6.5, if a sprite is parented to another, the parent sprite is
rotated and after said rotation, the child sprite's Scale is set (or its
sprite is changed, or SetAnimation is used) then the sprite's rotation
will be reset.

The variable "noscalerotationbug" has been introduced. Much like
"autolinebreak", if this variable is set to true in the Encounter script,
then the aforemented bug will be fixed, preventing the reset of the
rotation in this specific case.

This variable has been added to keep this version's faithfulness
regarding CYF 0.6.5, to make sure encounters from that version work the
very same in 0.6.5 and 0.6.6.