Delayed sprite deletion for dusting objects' sake
Now sprite objects are deleted a little bit later, allowing for Dust()
commands to be run on all sprites of a hierarchy if the root of said
hierarchy has been dusted first.

The actual object deletion code is run using the function LateUpdate(),
which is run after all Update() functions in the engine. This allows for
the current Lua frame to be performed and for all dusting instances to
be spawned before all sprites are properly deleted.